Fixes some links
A lot of links point to an old version of the web. This commits catchs up many
of then and links to the correct ones.

The [link](https://jekyllrb.com/docs/liquid/tags/#link) feature of jekyll is
used to done it in a consistent way.

Some markdown reformating is also done